Created Menu
The Photon Pun System is working and connects to the AUS servers

Also the main menu is basically finished with all the buttons and I just need to link them to other scenes to finish it.

Last Year we did a moving background so we are thinking of implementing is again this year. This will happen later when the map is finished.